refactor: removed avatar and separated checkbox
