+++ packages/checkbox/.gitignore
@@ -1,0 +1,5 @@
+*.log
+.DS_Store
+node_modules
+.cache
+dist
