+++ packages/checkbox/package.json
@@ -1,0 +1,50 @@
+{
+  "version": "0.1.0",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --passWithNoTests",
+    "lint": "tsdx lint",
+    "size": "size-limit",
+    "analyze": "size-limit --why",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
+  },
+  "peerDependencies": {
+    "react": ">=16"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  },
+  "name": "@highoutput/hds-checkbox",
+  "author": "John Christopher Go",
+  "module": "dist/hds-checkbox.esm.js",
+  "size-limit": [
+    {
+      "path": "dist/hds-checkbox.cjs.production.min.js",
+      "limit": "10 KB"
+    },
+    {
+      "path": "dist/hds-checkbox.esm.js",
+      "limit": "10 KB"
+    }
+  ]
+}
