+++ packages/checkbox/src/index.tsx
@@ -1,0 +1,8 @@
+import Checkbox, {
+  CheckProps,
+  CheckTypeProps,
+  CircleIcon,
+  RadioTypeProps,
+} from './checkbox';
+
+export { CheckProps, CheckTypeProps, Checkbox, CircleIcon, RadioTypeProps };
