+++ packages/checkbox/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "src/examples/**/*", "**/*.spec.*", "**/*.test.*"]
+}
